backend/requirements.txt
flask
flask-cors
torch
torchvision
transformers
librosa
soundfile
opencv-python
Pillow
scikit-learn
pandas
tabulate
pyspellchecker
joblib
numpy